requirements.txt
certifi>=2024.0.0
psycopg2-binary==2.9.11
pymongo==4.15.5
python-dotenv==1.2.2
requests==2.32.5
neo4j==6.1.0

